requirements.txt
requests
beautifulsoup4
lxml
azure-communication-email
azure-identity
azure-keyvault-secrets